pyproject.toml
[tool.poetry]
name = "nonebot_plugin_ayaka_games"
version = "0.0.1"
description = "a pack of textual game on QQ via nonebot-plugin-ayaka"
authors = ["Su <wxlxy316@163.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/bridgeL/nonebot-plugin-ayaka-games"
packages = [
    { include = "ayaka_games" },
]


[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0b5"
nonebot-adapter-onebot = "^2.1.3"
nonebot-plugin-ayaka = "^0.2.18"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
